// src/app/api/model/partyListCardModel.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { CandidateCardModel } from './candidateCardModel';


export interface PartyListCardModel { 
    id?: number;
    name?: string;
    imageUrl?: string;
    candidateModels?: Array<CandidateCardModel>;
}