// src/app/api/model/candidateCardModel.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface CandidateCardModel { 
    id?: number;
    firstName?: string;
    lastName?: string;
    counter?: number;
    mayor?: boolean;
}